Refactoring for better code organization
