+++ e2e/home.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 
-test('home page', async ({ page }) => {
+test('/ page', async ({ page }) => {
   await page.goto('http://localhost:3000');
   await expect(page).toHaveTitle(/BimaAdi/);
