+++ src/app/not-found.tsx
@@ -1,5 +1,5 @@
-import NotFound from "@/components/NotFound"
-import Main from "@/layouts/Main"
+import NotFound from "@/client/components/NotFound"
+import Main from "@/client/layouts/Main"
 
 export default function Custom404() {
     return (
