+++ src/client/components/Hero.tsx
