+++ src/client/components/OnProgress.tsx
@@ -12,7 +12,7 @@ const OnProgress = () => {
                     height: "60px",
                 }} />
                 <div className="w-2 bg-white h-full mx-5"></div>
-                <h3 className="text-3xl md:text-5xl">On Progress</h3>
+                <h3 id="onProgress" className="text-3xl md:text-5xl">On Progress</h3>
             </div>
         </div>
     )
