+++ tailwind.config.js
@@ -6,7 +6,7 @@ module.exports = {
     './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
     './src/components/**/*.{js,ts,jsx,tsx,mdx}',
     './src/app/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/layouts/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/client/**/*.{js,ts,jsx,tsx,mdx}',
   ],
   theme: {
     extend: {
